+++ src/Transactions/Transactions.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 import {
   Route,
   Switch,
@@ -7,13 +7,27 @@ import {
 import ReactRouterPropTypes from 'react-router-prop-types';
 
 import TransactionsList from './TransactionsList';
+import {
+  BUDGET_ROUTE,
+  BUDGET_VIEW_ROUTE,
+  FUNDS_ROUTE,
+} from '../common/const';
 
-const Transactions = ({ match }) => {
+const Transactions = ({ match, history }) => {
+  const closePane = useCallback(
+    () => {
+      return match.params.id
+        ? history.push(`${FUNDS_ROUTE}/view/${match.params.id}`)
+        : history.push(`${BUDGET_ROUTE}${match.params.budgetId}${BUDGET_VIEW_ROUTE}`);
+    },
+    [history, match.params],
+  );
+
   return (
     <Switch>
       <Route
         path={match.path}
-        component={TransactionsList}
+        render={() => <TransactionsList closePane={closePane} />}
       />
     </Switch>
   );
@@ -21,6 +35,7 @@ const Transactions = ({ match }) => {
 
 Transactions.propTypes = {
   match: ReactRouterPropTypes.match.isRequired,
+  history: ReactRouterPropTypes.history.isRequired,
 };
 
 export default withRouter(Transactions);
