+++ src/components/Main/Main.js
@@ -11,7 +11,6 @@ import { Ledger } from '../../Ledger';
 import Funds from '../../Funds';
 import Budget from '../Budget/Budget';
 import Groups from '../../Groups';
-import Transactions from '../../Transactions';
 import { FiscalYears } from '../../FiscalYears';
 
 const Main = ({ match }) => {
@@ -37,10 +36,6 @@ const Main = ({ match }) => {
         <Route
           path={FISCAL_YEAR_ROUTE}
           component={FiscalYears}
-        />
-        <Route
-          path={`${match.path}/:budgetId/transactions`}
-          component={Transactions}
         />
         <Redirect exact from={`${match.path}`} to={`${match.path}/ledger`} />
       </Switch>
